rename alertGenerator.py -> alertDataGenerator.py
