+++ python/lsst/sims/catUtils/utils/__init__.py
@@ -5,5 +5,5 @@ from .DBobjectTestUtils import *
 from .CatalogTestUtils import *
 from .LightCurveGenerator import *
 from .SNIaLightCurveGenerator import *
-from .alertGenerator import *
+from .alertDataGenerator import *
 from .avroAlertGenerator import *
